Update README.md
The module k8s_version in terraform-intersight-iks/main.tf seems to call var.versionPolicy.versionName.  So in this readme defining the version with iksVersionName would not allow terraform the find the moid of the desired versions.  Not sure if the problem is in this readme file or the module, but this changed seemed to work for my one use case.